refactor(persistence): ADR-0035 — relocate ShippedAuditRecord + verdict-id fns (enabling slice 3)
Two small root-crate couplings in verifier_store, both relocated to existing lean
leaves (no new crate, no lockfile churn):

- ShippedAuditRecord (crate::audit_shipper) -> kirra-core. It is the off-box
  SHIPPED-record wire type: ShippedAuditRecord + verify_shipped_chain are the
  INDEPENDENT re-verifier that must work with NO source DB, so the type must NOT
  live in the rusqlite-heavy persistence crate. kirra-core (lean, serde) is the
  right home for a wire type shared by the persistence producer and the off-box
  consumer. audit_shipper re-exports it; its recompute_hash inherent impl becomes
  a free fn calling kirra_audit_hash directly (the record type is now external) —
  which also freed audit_shipper of its own crate::audit_chain dependency.

- mint_verdict_id / is_valid_verdict_id (crate::verdicts) -> kirra-audit-hash.
  They are content-addressed SHA-256 audit ids (same sha2+hex deps), so they sit
  naturally with the audit hash primitives. verdicts re-exports them.

verifier_store's audit.rs now names kirra_core::ShippedAuditRecord and
kirra_audit_hash::is_valid_verdict_id directly. (KeyRegistry, the third item in
the ADR's list, turned out to be doc-links only — no code coupling.)

Verified: cargo test --lib 760, audit_shipper module 16 (verify_shipped_chain /
sinks / ship_and_advance — confirms the recompute_hash + relocation are
behaviour-preserving), power-loss drill 3, kirra-core 184, fmt, guardrails,
orphan-cores.

Remaining before the wholesale kirra-persistence move: repoint the store's
ROOT-shim paths to the leaf crates directly (crate::verifier -> kirra_core,
crate::federation* -> kirra_fleet_types, crate::ota_campaign -> kirra_ota_campaign,
crate::fabric::* -> kirra_fabric_types) + one test-only gateway coupling.
